Update
Code cleanup for compilation with Clang and MSVC
Add user service layer
Add an option to disable default key behavior
Disable notifier surface from rendering
Instanciable Tracer
Refactor Tracer initilization with logger to use std location
Update primary services identification
Improve Window fullscreen management
Adds C++ constants instead of macros
Add the visibility check per surfaces for input in overlay manager.
Enable Wayland support.
Update some constexpr usage.
GTK3 removed.
Add dialog box cancelable state.
Crash fixed on some situation.
Shader generators flexibility improved.
Renderable instance code improved.
Semaphores and fences improved in render service.
Compatible with PCH compiler feature.
Imrove thread sleep code.
Tracer update.

+++ src/Graphics/Material/Interface.hpp
@@ -403,7 +403,7 @@ namespace EmEn::Graphics::Material
 			 * @return size_t
 			 */
 			[[nodiscard]]
-			virtual size_t frameIndexAt (uint32_t sceneTime) const noexcept = 0;
+			virtual uint32_t frameIndexAt (uint32_t sceneTime) const noexcept = 0;
 
 			/**
 			 * @brief Enables a blending and set the desired mode.
